requirements.txt
streamlit
opencv-python
transformers
torch
pillow
moviepy
gtts
scenedetect[opencv]